split parser into multiple files
