+++ src/main/java/com/github/sybila/huctl/Assignment.kt
@@ -1,0 +1,13 @@
+package com.github.sybila.huctl
+
+/**
+ * Abstraction over assignment expressions in the HUCTLp file.
+ *
+ * It remembers basic info such as flag and location for future processing.
+ */
+internal data class Assignment<out T: Any>(
+        val name: String,
+        val item: T,
+        val location: String,
+        val flagged: Boolean
+)
