+++ src/main/java/com/github/sybila/huctl/Extensions.kt
@@ -69,113 +69,3 @@ fun Formula.mapLeafs(
 ): Formula {
     return this.fold(atom, Unary<*>::copy, Binary<*>::copy)
 }
-
-// -- Construction Utils --
-
-// Expressions
-
-fun String.asVariable(): Expression.Variable = Expression.Variable(this)
-fun Double.asConstant(): Expression.Constant = Expression.Constant(this)
-
-infix operator fun Expression.plus(other: Expression) = Expression.Add(this, other)
-infix operator fun Expression.minus(other: Expression) = Expression.Sub(this, other)
-infix operator fun Expression.times(other: Expression) = Expression.Mul(this, other)
-infix operator fun Expression.div(other: Expression) = Expression.Div(this, other)
-
-infix fun Expression.gt(other: Expression): Numeric = Numeric(this, CompareOp.GT, other)
-infix fun Expression.ge(other: Expression): Numeric = Numeric(this, CompareOp.GE, other)
-infix fun Expression.eq(other: Expression): Numeric = Numeric(this, CompareOp.EQ, other)
-infix fun Expression.neq(other: Expression): Numeric = Numeric(this, CompareOp.NEQ, other)
-infix fun Expression.le(other: Expression): Numeric = Numeric(this, CompareOp.LE, other)
-infix fun Expression.lt(other: Expression): Numeric = Numeric(this, CompareOp.LT, other)
-
-// Propositions
-
-val True = Formula.True
-val False = Formula.False
-
-fun String.positiveIn() = Transition(this, Flow.IN, Direction.POSITIVE)
-fun String.positiveOut() = Transition(this, Flow.OUT, Direction.POSITIVE)
-fun String.negativeIn() = Transition(this, Flow.IN, Direction.NEGATIVE)
-fun String.negativeOut() = Transition(this, Flow.OUT, Direction.NEGATIVE)
-fun String.asReference() = Reference(this)
-
-// Flow propositions
-
-fun String.increase(): DirFormula = DirFormula.Proposition(this, Direction.POSITIVE)
-fun String.decrease(): DirFormula = DirFormula.Proposition(this, Direction.NEGATIVE)
-
-// Flow formulas
-
-fun not(inner: DirFormula): DirFormula = DirFormula.Not(inner)
-
-infix fun DirFormula.and(right: DirFormula): DirFormula = DirFormula.And(this, right)
-infix fun DirFormula.or(right: DirFormula): DirFormula = DirFormula.Or(this, right)
-infix fun DirFormula.implies(right: DirFormula): DirFormula = DirFormula.Implies(this, right)
-infix fun DirFormula.equal(right: DirFormula): DirFormula = DirFormula.Equals(this, right)
-
-
-// Formulas
-
-fun not(inner: Formula): Formula = Not(inner)
-
-//Future
-fun EF(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Future(E, inner, dir)
-fun pastEF(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Future(pE, inner, dir)
-fun AF(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Future(A, inner, dir)
-fun pastAF(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Future(pA, inner, dir)
-
-//WeakFuture
-fun weakEF(inner: Formula, dir: DirFormula = DirFormula.True): Formula = WeakFuture(E, inner, dir)
-fun weakAF(inner: Formula, dir: DirFormula = DirFormula.True): Formula = WeakFuture(A, inner, dir)
-fun pastWeakEF(inner: Formula, dir: DirFormula = DirFormula.True): Formula = WeakFuture(pE, inner, dir)
-fun pastWeakAF(inner: Formula, dir: DirFormula = DirFormula.True): Formula = WeakFuture(pA, inner, dir)
-
-//Next
-fun EX(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Next(E, inner, dir)
-fun pastEX(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Next(pE, inner, dir)
-fun AX(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Next(A, inner, dir)
-fun pastAX(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Next(pA, inner, dir)
-
-//WeakNext
-fun weakEX(inner: Formula, dir: DirFormula = DirFormula.True): Formula = WeakNext(E, inner, dir)
-fun pastWeakEX(inner: Formula, dir: DirFormula = DirFormula.True): Formula = WeakNext(pE, inner, dir)
-fun weakAX(inner: Formula, dir: DirFormula = DirFormula.True): Formula = WeakNext(A, inner, dir)
-fun pastWeakAX(inner: Formula, dir: DirFormula = DirFormula.True): Formula = WeakNext(pA, inner, dir)
-
-//Globally
-fun EG(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Globally(E, inner, dir)
-fun pastEG(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Globally(pE, inner, dir)
-fun AG(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Globally(A, inner, dir)
-fun pastAG(inner: Formula, dir: DirFormula = DirFormula.True): Formula = Globally(pA, inner, dir)
-
-//Until
-infix fun Formula.EU(reach: Formula): Formula
-        = Until(E, this, reach, DirFormula.True)
-fun Formula.EU(reach: Formula, dir: DirFormula = DirFormula.True): Formula
-        = Until(E, this, reach, dir)
-infix fun Formula.AU(reach: Formula): Formula
-        = Until(A, this, reach, DirFormula.True)
-fun Formula.AU(reach: Formula, dir: DirFormula = DirFormula.True): Formula
-        = Until(A, this, reach, dir)
-infix fun Formula.pastEU(reach: Formula): Formula
-        = Until(pE, this, reach, DirFormula.True)
-fun Formula.pastEU(reach: Formula, dir: DirFormula = DirFormula.True): Formula
-        = Until(pE, this, reach, dir)
-infix fun Formula.pastAU(reach: Formula): Formula
-        = Until(pA, this, reach, DirFormula.True)
-fun Formula.pastAU(reach: Formula, dir: DirFormula = DirFormula.True): Formula
-        = Until(pA, this, reach, dir)
-
-//Hybrid and first order
-
-fun at(name: String, inner: Formula) : Formula = Formula.At(name, inner)
-fun bind(name: String, inner: Formula) : Formula = Formula.Bind(name, inner)
-fun forall(name: String, bound: Formula, inner: Formula) : Formula = ForAll(name, bound, inner)
-fun exists(name: String, bound: Formula, inner: Formula) : Formula = Exists(name, bound, inner)
-
-//Bool
-infix fun Formula.and(other: Formula): Formula = And(this, other)
-infix fun Formula.or(other: Formula): Formula = Or(this, other)
-infix fun Formula.implies(other: Formula): Formula = Implies(this, other)
-infix fun Formula.equal(other: Formula): Formula = Equals(this, other)
